# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Symphony Orchestrator API
  version: 1.0.0
  description: Symphony Orchestrator API for Vertex Workbench provisioning and quota management
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects/{projectId}/self/usage:
    get:
      tags: [Usage]
      summary: Get my current Workbench usage and remaining quota (computed via Cloud Asset Inventory)
      operationId: getSelfUsage
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/groupId"
      responses:
        "200":
          description: Usage snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserQuotaUsage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
      security:
        - bearerAuth: [workbench.user]

  # -------------------------
  # Quotas (Manager) - Group-scoped user quotas (distinct groups in UI)
  # -------------------------
  /v1/projects/{projectId}/managed-groups/{groupId}/quotas/users/{userId}:
    put:
      tags: [Quotas]
      summary: Upsert a user's hard quota within a specific managed user-group scope
      description: |
        Group-scoped to preserve distinct group views in the UI.
        Caller must manage {groupId}, and {userId} must be a member of that groupId.
      operationId: upsertUserQuotaInGroup
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserQuotaSpec" }
      responses:
        "200":
          description: Upserted quota
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserQuota" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not authorized for this group/user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.manager]

    get:
      tags: [Quotas]
      summary: Get a user's quota within a managed user-group scope
      operationId: getUserQuotaInGroup
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User quota lookup result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserQuotaLookupResult" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not authorized for this group/user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.manager]

    delete:
      tags: [Quotas]
      summary: Delete a user's quota (revert to default) within a managed user-group scope
      operationId: deleteUserQuotaInGroup
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/userId"
      responses:
        "204":
          description: Deleted (idempotent — also returns 204 if no explicit quota existed)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not authorized for this group/user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.manager]
...
  /v1/projects/{projectId}/managed-groups/{groupId}/quotas/users/{userId}/audit/latest:
    get:
      tags: [Quotas]
      summary: Get latest quota change (who/when/what) for a user within a managed group context
      operationId: getLatestUserQuotaAudit
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: Latest audit record (or NONE if no changes exist)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LatestQuotaAuditResult" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Not authorized for this group/user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.manager]

  # -------------------------
  # Default quota (Manager) - project-wide fallback
  # -------------------------
  /v1/projects/{projectId}/quotas/default:
    get:
      tags: [Quotas]
      summary: Get project default quota (manager)
      operationId: getDefaultQuota
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Default quota
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DefaultQuota" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No project default quota has been set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.manager]

    put:
      tags: [Quotas]
      summary: Upsert project default quota (manager)
      operationId: upsertDefaultQuota
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserQuotaSpec" }
      responses:
        "200":
          description: Updated default quota
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DefaultQuota" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      security:
        - bearerAuth: [workbench.manager]

    delete:
      tags: [Quotas]
      summary: Delete project default quota (manager)
      operationId: deleteDefaultQuota
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No project default quota exists to delete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.manager]
...
components:
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
    groupId:
      name: groupId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/GroupId" }
    userId:
      name: userId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/BankUserId" }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      # NOTE: OpenAPI scopes (e.g. bearerAuth: [workbench.manager]) are only enforced
      # for oauth2/openIdConnect scheme types. For http bearer, the scope array is
      # ignored by validators and tooling — it is used here as intent documentation only.
      # Actual scope enforcement is performed server-side via Spring Security @PreAuthorize.
  schemas:
...
    BankUserId:
      type: string
      description: 1bankid used in Workbench labels.user
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9._-]+$'
    GroupId:
      type: string
      description: Unique directory group identifier (GUID/DN)
      minLength: 1
      maxLength: 128
...
    UserQuotaSpec:
      type: object
      required: [maxInstances, maxTotalVcpu, allowedMachineTypes]
      additionalProperties: false
      properties:
        maxInstances:
          type: integer
          minimum: 0
          description: Manager-configured limit; system cap 3 still applies at enforcement.
        maxTotalVcpu:
          type: integer
          minimum: 0
        allowedMachineTypes:
          type: array
          items: { type: string }
          description: |
            Machine type allowlist. An empty array means no restriction (any machine type is permitted).
            Must contain at least one entry to actively restrict choices.
        maxBootDiskGb:
          type: integer
          minimum: 50
          default: 200
        allowedZones:
          type: array
          items: { type: string }
          description: Zone allowlist. An empty array means no zone restriction.
    UserQuota:
      type: object
      required: [projectId, groupId, userId, spec, updatedAt, updatedBy]
      properties:
        projectId: { type: string }
        groupId:
          allOf:
            - $ref: "#/components/schemas/GroupId"
          description: The managed user-group context the manager used when editing (keeps groups distinct).
        userId: { $ref: "#/components/schemas/BankUserId" }
        spec: { $ref: "#/components/schemas/UserQuotaSpec" }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string, description: Manager principal (1bankid or subject) }
    DefaultQuota:
      type: object
      required: [projectId, spec, updatedAt, updatedBy]
      properties:
        projectId: { type: string }
        spec: { $ref: "#/components/schemas/UserQuotaSpec" }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
    UserQuotaLookupResult:
      type: object
      required: [projectId, groupId, userId, status, quota, effectiveSource]
      properties:
        projectId: { type: string }
        groupId: { $ref: "#/components/schemas/GroupId" }
        userId: { $ref: "#/components/schemas/BankUserId" }
        status:
          type: string
          enum: [EXPLICIT, NONE]
        quota:
          oneOf:
            - $ref: "#/components/schemas/UserQuota"
            - type: "null"
        effectiveSource:
          type: string
          enum: [USER_QUOTA, DEFAULT_QUOTA, GLOBAL_DEFAULT]
          description: |
            Where the effective quota limit comes from:
            - USER_QUOTA: an explicit per-user quota exists (status=EXPLICIT)
            - DEFAULT_QUOTA: no explicit user quota; falling back to the project default set via PUT /quotas/default
            - GLOBAL_DEFAULT: no explicit user quota and no project default; system hardcoded cap of 3 instances applies
        note: { type: string }
...
    SystemCaps:
      type: object
      required: [maxInstancesPerUser]
      properties:
        maxInstancesPerUser:
          type: integer
          const: 3
          description: Hard platform cap enforced regardless of manager quota.
    EffectiveLimits:
      type: object
      required: [maxInstances, maxTotalVcpu, allowedMachineTypes]
      properties:
        maxInstances:
          type: integer
          minimum: 0
          description: min(systemCaps.maxInstancesPerUser=3, manager/default quota maxInstances)
        maxTotalVcpu:
          type: integer
          minimum: 0
        allowedMachineTypes:
          type: array
          items: { type: string }
        maxBootDiskGb:
          type: integer
          minimum: 50
        allowedZones:
          type: array
          items: { type: string }
    UsageTotals:
      type: object
      required: [instances, totalVcpu]
      properties:
        instances: { type: integer, minimum: 0 }
        totalVcpu: { type: integer, minimum: 0 }
    UsageComputation:
      type: object
      required: [computedFrom, assetType, query]
      properties:
        computedFrom:
          type: string
          enum: [CLOUD_ASSET_INVENTORY]
        assetType:
          type: string
          const: compute.googleapis.com/Instance
        query:
          type: string
          description: CAI query used.
          example: "labels.notebooks-product:* AND labels.user:u12345"
    UserQuotaUsage:
      type: object
      required: [projectId, userId, computedAt, computation, systemCaps, effectiveLimits, usage, remaining]
      properties:
        projectId: { type: string }
        userId: { $ref: "#/components/schemas/BankUserId" }
        computedAt: { type: string, format: date-time }
        computation: { $ref: "#/components/schemas/UsageComputation" }
        systemCaps: { $ref: "#/components/schemas/SystemCaps" }
        effectiveLimits: { $ref: "#/components/schemas/EffectiveLimits" }
        usage: { $ref: "#/components/schemas/UsageTotals" }
        remaining: { $ref: "#/components/schemas/UsageTotals" }
...
    LatestQuotaAuditResult:
      type: object
      required: [projectId, groupId, userId, status, latest]
      properties:
        projectId: { type: string }
        groupId: { $ref: "#/components/schemas/GroupId" }
        userId: { $ref: "#/components/schemas/BankUserId" }
        status:
          type: string
          enum: [PRESENT, NONE]
        latest:
          oneOf:
            - $ref: "#/components/schemas/QuotaAuditEvent"
            - type: "null"
    QuotaAuditEvent:
      type: object
      required: [eventId, action, at, actor, context]
      properties:
        eventId: { type: string }
        action:
          type: string
          enum: [UPSERT, DELETE]
        at:
          type: string
          format: date-time
        actor:
          $ref: "#/components/schemas/AuditActor"
        context:
          $ref: "#/components/schemas/AuditContext"
        oldSpec:
          oneOf:
            - $ref: "#/components/schemas/UserQuotaSpec"
            - type: "null"
        newSpec:
          oneOf:
            - $ref: "#/components/schemas/UserQuotaSpec"
            - type: "null"
    AuditActor:
      type: object
      required: [principal]
      properties:
        principal:
          type: string
          description: Manager identity (JWT subject / 1bankid / email depending on your auth)
        displayName: { type: string }
        email: { type: string }
    AuditContext:
      type: object
      required: [groupId]
      properties:
        groupId:
          allOf:
            - $ref: "#/components/schemas/GroupId"
          description: Managed user-group context used for the change (distinct group UI)
        managerGroupId:
          oneOf:
            - $ref: "#/components/schemas/GroupId"
            - type: "null"
          description: |
            The *_Managers group the actor belongs to, derived from the *_Managers↔*_Users naming convention.
            Null only when the change was made by a system/admin process outside the manager group pairing flow.

    # -------------------------
    # Workbench instance models
    # -------------------------
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string }
        requestId: { type: string, description: Correlation ID for tracing this error in server logs }